Extend ec_ctrl syscall to set and get vCPU state
Op number 6 reads the vCPU state into the UTCB and reads the vCPU's
FPU state if needed.

Op number 7 updates the vCPU state from the UTCB and updates the vCPU's
FPU state from the current FPU state if needed.

The syscall must be performed by a thread on the same pCPU as the vCPU.

Ref genodelabs/genode#4968