lapic: save/restore AP code on demand
Using a segment to reserve the AP code page does not work if the region
overlaps with other non-memory types, e.g. bender checks for the overlapping
and denies to boot on such machines. So, partially revert the segment
reservation of "Add segment to reserve AP startup page".

Instead apply the idea of Cyberus version of NOVA, to place the AP boot code
later during boot and save and restore the memory before/after use.